Se mejoro Informe.java
Se agrego en el informe la cantidad y porcentaje de Asistentes

Se mejoro la impresion de los Asistentes, Atrasos, Ausentes y las Notas

+++ tarea2/src/main/java/Clases/Reunion.java
@@ -149,7 +149,7 @@ abstract public class Reunion{
      * @return Porcentaje de asistencia de la reunion.
      */
     public float obtenerPorcentajeAsistencia(){
-        return (obtenerTotalAsistencia()/(obtenerTotalAsistencia() + obtenerAusencias().size()))*100;
+        return ((float)obtenerTotalAsistencia()/(obtenerTotalAsistencia() + obtenerAusencias().size()))*100;
     }
